feat(corr_plots.R): add correlation statistics to plots for enhanced data interpretation
Add `stat_correlation` to display correlation statistics on plots.
This provides additional context and insights into the relationships
depicted in the plots.

chore(targets.R): add ggpmisc library for correlation statistics

Include `ggpmisc` library to support the addition of correlation
statistics in plots, enhancing the analytical capabilities of the
visualizations.